fix 1

+++ users/models/usersAccessDataService.js
@@ -5,7 +5,7 @@ const { comparePassword } = require("../helpers/bcrypt");
 const { generateAuthToken } = require("../../auth/Providers/jwt");
 const { handleBadRequest } = require("../../utils/handleErrors");
 
-const registerUser = async normalizedUser => {
+const registerUser = async (normalizedUser) => {
   if (DB === "MONGODB") {
     try {
       const { email } = normalizedUser;
@@ -59,7 +59,7 @@ const getUsers = async () => {
   return Promise.resolve("get users not in mongodb");
 };
 
-const getUser = async userId => {
+const getUser = async (userId) => {
   if (DB === "MONGODB") {
     try {
       let user = await User.findById(userId, {
@@ -91,7 +91,7 @@ const updateUser = async (userId, normalizedUser) => {
   return Promise.resolve("card update not in mongodb");
 };
 
-const changeUserBusinessStatus = async userId => {
+const changeUserBusinessStatus = async (userId) => {
   if (DB === "MONGODB") {
     try {
       const pipeline = [{ $set: { isBusiness: { $not: "$isBusiness" } } }];
@@ -113,10 +113,14 @@ const changeUserBusinessStatus = async userId => {
   return Promise.resolve("card liked not in mongodb");
 };
 
-const deleteUser = async userId => {
+const deleteUser = async (userId) => {
   if (DB === "MONGODB") {
     try {
-      const user = await User.findByIdAndDelete(userId, {
+      let user = await User.findById(userId);
+      if (user.isAdmin) {
+        throw new Error("Could not delete admin user");
+      }
+      user = await User.findByIdAndDelete(userId, {
         password: 0,
         __v: 0,
       });
